flask/requirements.txt
asttokens==2.1.0
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.3
decorator==5.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.1/en_core_web_md-3.4.1-py3-none-any.whl
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
fonttools==4.38.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==5.0.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pathy==0.10.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
preshed==3.0.8
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyPdf==1.13
PyPDF2==2.11.2
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0.post1
smart-open==5.2.1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
stack-data==0.6.1
tesseract==0.1.3
thinc==8.1.5
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
